add logout button
- minor changes to signIn flow.
