+++ ui/src/app/auth/sign-in/sign-in.component.html
@@ -2,7 +2,7 @@
         #signInForm="ngForm">
   
     <div class="container">
-      <label for="userName"><b>Email address</b></label>
+      <label for="userName"><b>Username</b></label>
       <input 
       name="userName"
       ngModel
