+++ ui/src/app/auth/sign-up/sign-up.component.html
@@ -23,7 +23,7 @@
     
         <label for="email"><b>Mobile No.</b></label>
         <input type="text" ngModel #mobileNoInput="ngModel" [(ngModel)]="mobileNo"
-        placeholder="Enter mobile no" name="mobileNo" id="mobileNo" required>
+        placeholder="Enter mobile no" name="mobileNo" id="mobileNo">
     
         <label for="psw"><b>Password</b></label>
         <input type="password"
