+++ ui/src/app/view/dashboard/dashboard.component.ts
@@ -38,15 +38,4 @@ export class DashboardComponent {
 
   transportCallSelectHandler = ($transportCall: TransportCall) => {this.transportCallSelected = $transportCall;};
 
-  onLogout(): void {
-    let poolData = {
-      UserPoolId: environment.cognitoUserPoolId,
-      ClientId: environment.cognitoAppClientId
-    };
-    let userPool = new CognitoUserPool(poolData);
-    let cognitoUser = userPool.getCurrentUser();
-    cognitoUser?.signOut();
-    this.router.navigate(["signin"])
-  }
-
 }
