+++ ui/src/app/view/header/header.component.html
@@ -7,7 +7,7 @@
     <p-menu #menu [popup]="true" [model]="helpMenu"></p-menu>
   </div>
   <div class="p-col-1">
-
+    <button (click)="onLogout()">Log out</button>
 
   </div>
